Rename internal/resource to xresource
To avoid a package name conflict with the crossplane-runtime resource
package. xresource builds on resource - it adds things that pertain only
to composition and are consumed only by core Crossplane.
