+++ internal/xresource/unstructured/generate.go
